+++ src/main/java/com/trxyzng/trung/authentication/shared/beanConfigs/SqlConfig.java
@@ -1,4 +1,4 @@
-package com.trxyzng.trung.comment;
+package com.trxyzng.trung.authentication.shared.beanConfigs;
 
 import org.springframework.boot.autoconfigure.jdbc.DataSourceProperties;
 import org.springframework.boot.context.properties.ConfigurationProperties;
